+++ src/main/java/onboarding/problem4/domain/Sentence.java
@@ -1,6 +1,6 @@
 package onboarding.problem4.domain;
 
-import onboarding.problem4.util.AlphabetConvertor;
+import onboarding.problem4.utils.AlphabetConvertor;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
